Update immigrant_lexicon_en.txt
The list blends race and ethnicity and rural/urban residence  with migration, which might produce analytically biased results
Many important migrant expelling countries are not in the original list
Some words are country specific. For example, black might have a migration-related connotation in the UK but not in the US. Maybe it would be a good idea to make country-specific lists?